Check that the image format is correct.
I want to deal only with JPEG and PNG files. After opening the image,
I check that the "format" attribute is one of those. If not, the
CLI exits prompting an error to the user.

Moreover, I create some variable to store format and filename for
future reference. This is done because, when the current image will
be modified, those informations are removed (for example when
resize is applied).