Complete resize function.
The CLI provide a quick resize function if the --resize option is
passed. The number provided should be the width of the new image,
and the function will calculate the height to maintain proportions.

When resampling, the Image.LANCZOS filter provided by the Pillow library
is applied by deafault. This was chosen because, even if it is the less
perfomant filter, it has the best image quality output both in upscaling
and downscaling.