Complete saving process for both the image formats.
The conversion from a format to another happens only in the last
stage of the modification process. Moreover, while saving options are
passed to imporove image quality for web usage. In particular, optimization
is always applied to reduce image weight, and dpi is set to 72.

JPEG images are also saved in progressive mode, to improve web perfromance
while loading.